move proxy related config to proxy.go
