+++ config/proxy.go
@@ -1,0 +1,114 @@
+package config
+
+import (
+	"fmt"
+	"net/http"
+	"net/url"
+	"os"
+	"strconv"
+	"strings"
+
+	"github.com/go-ini/ini"
+)
+
+// getProxySettings returns a url.Url for the proxy configuration from datadog.conf, if available.
+// In the case of invalid settings an error is logged and nil is returned. If settings are missing,
+// meaning we don't want a proxy, then nil is returned with no error.
+func getProxySettings(m *ini.Section) (proxyFunc, error) {
+	var host string
+	scheme := "http"
+	if v := m.Key("proxy_host").MustString(""); v != "" {
+		// accept either http://myproxy.com or myproxy.com
+		if i := strings.Index(v, "://"); i != -1 {
+			// when available, parse the scheme from the url
+			scheme = v[0:i]
+			host = v[i+3:]
+		} else {
+			host = v
+		}
+	}
+
+	if host == "" {
+		return nil, nil
+	}
+
+	port := defaultProxyPort
+	if v := m.Key("proxy_port").MustInt(-1); v != -1 {
+		port = v
+	}
+	var user, password string
+	if v := m.Key("proxy_user").MustString(""); v != "" {
+		user = v
+	}
+	if v := m.Key("proxy_password").MustString(""); v != "" {
+		password = v
+	}
+	return constructProxy(host, scheme, port, user, password)
+}
+
+// proxyFromEnv parses out the proxy configuration from the ENV variables in a
+// similar way to getProxySettings and, if enough values are available, returns
+// a new proxy URL value. If the environment is not set for this then the
+// `defaultVal` is returned.
+func proxyFromEnv(defaultVal proxyFunc) (proxyFunc, error) {
+	var host string
+	scheme := "http"
+	if v := os.Getenv("PROXY_HOST"); v != "" {
+		// accept either http://myproxy.com or myproxy.com
+		if i := strings.Index(v, "://"); i != -1 {
+			// when available, parse the scheme from the url
+			scheme = v[0:i]
+			host = v[i+3:]
+		} else {
+			host = v
+		}
+	}
+
+	if host == "" {
+		return defaultVal, nil
+	}
+
+	port := defaultProxyPort
+	if v := os.Getenv("PROXY_PORT"); v != "" {
+		port, _ = strconv.Atoi(v)
+	}
+	var user, password string
+	if v := os.Getenv("PROXY_USER"); v != "" {
+		user = v
+	}
+	if v := os.Getenv("PROXY_PASSWORD"); v != "" {
+		password = v
+	}
+
+	return constructProxy(host, scheme, port, user, password)
+}
+
+// constructProxy constructs a *url.Url for a proxy given the parts of a
+// Note that we assume we have at least a non-empty host for this call but
+// all other values can be their defaults (empty string or 0).
+func constructProxy(host, scheme string, port int, user, password string) (proxyFunc, error) {
+	var userpass *url.Userinfo
+	if user != "" {
+		if password != "" {
+			userpass = url.UserPassword(user, password)
+		} else {
+			userpass = url.User(user)
+		}
+	}
+
+	var path string
+	if userpass != nil {
+		path = fmt.Sprintf("%s@%s:%v", userpass.String(), host, port)
+	} else {
+		path = fmt.Sprintf("%s:%v", host, port)
+	}
+	if scheme != "" {
+		path = fmt.Sprintf("%s://%s", scheme, path)
+	}
+
+	u, err := url.Parse(path)
+	if err != nil {
+		return nil, err
+	}
+	return http.ProxyURL(u), nil
+}
